requirements.txt
pandas==2.2.3
pyyaml==6.0.2
